services code refactor
