+++ src/main/java/de/aittr/online_lessons/controllers/CourseController.java
@@ -2,7 +2,7 @@ package de.aittr.online_lessons.controllers;
 
 import de.aittr.online_lessons.controllers.api.CourseApi;
 import de.aittr.online_lessons.domain.dto.CourseDto;
-import de.aittr.online_lessons.services.jpa.CourseService;
+import de.aittr.online_lessons.services.CourseService;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
