+++ src/main/java/de/aittr/online_lessons/controllers/UserController.java
@@ -2,10 +2,7 @@ package de.aittr.online_lessons.controllers;
 
 import de.aittr.online_lessons.controllers.api.UserApi;
 import de.aittr.online_lessons.domain.dto.UserDto;
-import de.aittr.online_lessons.services.jpa.UserService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import jakarta.validation.Valid;
+import de.aittr.online_lessons.services.UserService;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
