+++ src/main/java/de/aittr/online_lessons/services/CourseService.java
@@ -1,4 +1,4 @@
-package de.aittr.online_lessons.services.jpa;
+package de.aittr.online_lessons.services;
 
 import de.aittr.online_lessons.domain.dto.CourseDto;
 import de.aittr.online_lessons.domain.jpa.Course;
@@ -6,7 +6,7 @@ import de.aittr.online_lessons.domain.jpa.User;
 import de.aittr.online_lessons.exception_handling.exceptions.CourseNotFoundException;
 import de.aittr.online_lessons.exception_handling.exceptions.CourseValidationException;
 import de.aittr.online_lessons.repositories.jpa.CourseRepository;
-import de.aittr.online_lessons.services.mapping.CourseMappingService;
+import de.aittr.online_lessons.mapping.CourseMappingService;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
